% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_io_points}
\alias{add_io_points}
\title{Add I/O Points to the Warehouse Arcs}
\usage{
add_io_points(arcs)
}
\arguments{
\item{arcs}{A data frame or matrix representing the arcs in the warehouse.}
}
\value{
A matrix of arcs with I/O points added.
}
\description{
This internal function adds I/O (input/output) points to the warehouse arcs.
}
\keyword{internal}
