% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{count_by_sequence}
\alias{count_by_sequence}
\title{Count Sequential Occurrences in a Vector}
\usage{
count_by_sequence(v)
}
\arguments{
\item{v}{A vector of values to be analyzed.}
}
\value{
A data frame with two columns: `id` representing the unique values and `count` representing the number of consecutive occurrences of each value.
}
\description{
This internal function counts the number of sequential occurrences of values in a vector and returns the results as a data frame.
}
\keyword{internal}
