% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{delete_empty_final_aisles}
\alias{delete_empty_final_aisles}
\title{Delete Empty Final Aisles from Arcs}
\usage{
delete_empty_final_aisles(arcs, warehouse_width_reduced)
}
\arguments{
\item{arcs}{A data frame or matrix representing the arcs in the warehouse.}

\item{warehouse_width_reduced}{The reduced width of the warehouse.}
}
\value{
A matrix of arcs with empty final aisles removed.
}
\description{
This internal function removes arcs corresponding to empty final aisles from the warehouse.
}
\keyword{internal}
