% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{edges_from_aisle}
\alias{edges_from_aisle}
\title{Retrieve Edges from an Aisle}
\usage{
edges_from_aisle(arcs, aisle)
}
\arguments{
\item{arcs}{A data frame or matrix containing arc data, where each row represents an edge with `x_from`, `x_to`, `y_from`, and `y_to` values.}

\item{aisle}{An integer representing the aisle number to filter the edges.}
}
\value{
A data frame of edges that correspond to the specified aisle, ordered by the `y_from` values.
}
\description{
This internal function retrieves the edges (connections) within a specific aisle from a set of arcs.
The edges are filtered based on the `x_from` and `x_to` values, and ordered by `y_from`.
}
\keyword{internal}
