% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sample_goods_and_locatons_scenario.R
\name{generate_sample_goods_and_locatons_scenario}
\alias{generate_sample_goods_and_locatons_scenario}
\title{Generate Sample Goods and Locations Scenario for a Warehouse}
\usage{
generate_sample_goods_and_locatons_scenario(
  warehouse_height = 25,
  warehouse_width = 40,
  warehouse_size = warehouse_height * warehouse_height,
  nr_goods = 10,
  max_nr_locations = rep(10, nr_goods),
  max_storage_time = rep(30, nr_goods),
  max_demand = rep(200, nr_goods),
  actual_demand = rep(100, nr_goods),
  full_packaging_amount = rep(20, nr_goods)
)
}
\arguments{
\item{warehouse_height}{An integer specifying the height of the warehouse (default is 25).}

\item{warehouse_width}{An integer specifying the width of the warehouse (default is 40).}

\item{warehouse_size}{The total size of the warehouse (height * width), default is calculated as `warehouse_height * warehouse_height`.}

\item{nr_goods}{An integer specifying the number of different goods/products in the warehouse (default is 10).}

\item{max_nr_locations}{A vector specifying the maximum number of locations each product can occupy (default is `rep(10, nr_goods)`).}

\item{max_storage_time}{A vector specifying the maximum storage time for each product (default is `rep(30, nr_goods)`).}

\item{max_demand}{A vector specifying the maximum demand for each product (default is `rep(200, nr_goods)`).}

\item{actual_demand}{A vector specifying the actual demand for each product (default is `rep(100, nr_goods)`).}

\item{full_packaging_amount}{A vector specifying the full packaging amount for each product (default is `rep(20, nr_goods)`).}
}
\value{
A data frame containing the generated scenario data with columns for:

- `product`: Product identifier.

- `location`: Location of the product in the warehouse.

- `storage_time`: Storage time of the product.

- `demand_ratio`: The ratio of demand relative to the actual demand.

- `full_packages_demand_ratio`: The ratio of full packages relative to the demand.

- `distance`: The calculated distance to the I/O point.

- `neighbors`: The number of neighboring products for each location.
}
\description{
This function generates a sample scenario of goods and their locations in a warehouse, along with related data such as storage time, demand ratio, and full packages demand ratio.
}
\details{
The function randomly assigns goods to locations in the warehouse and calculates various parameters such as storage time, demand ratio, and the full packages demand ratio for each product. It also calculates the number of neighboring products at each location and the distance from the location to the I/O point.
}
\examples{
scenario <- generate_sample_goods_and_locatons_scenario()
scenario2 <- generate_sample_goods_and_locatons_scenario(warehouse_height = 10,
warehouse_width = 10,nr_goods = 5)


}
\references{
Dmytrów, K. (2022). Analytical and simulation determination of order picking time in a low storage warehouse for shared storage systems. Operations Research and Decisions, 32(2), 34–51. \doi{10.37190/ord220203}
}
\author{
Krzysztof Dmytrów \email{krzysztof.dmytrow@usz.edu.pl} [aut] \href{https://orcid.org/0000-0001-7657-6063}{ORCID: 0000-0001-7657-6063}

Andrzej Dudek \email{andrzej.dudek@ue.wroc.pl} [aut, cre] \href{https://orcid.org/0000-0002-4943-8703}{ORCID: 0000-0002-4943-8703}
}
