% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parity_full_s}
\alias{parity_full_s}
\title{Check Parity of Sequences}
\usage{
parity_full_s(counted)
}
\arguments{
\item{counted}{A data frame created by `count_by_sequence` with columns `id` and `count`.}
}
\value{
A logical value indicating whether the sequence has valid parity (`TRUE`) or not (`FALSE`).
}
\description{
This internal function checks the parity (even or odd) of sequences of consecutive occurrences in a data frame created by `count_by_sequence`.
}
\keyword{internal}
